Added methods to the step defs
Copied the methods from the test file.  Test passes successfully.  I
noticed that when the test is ran that it wasn't logging out.  So I
tried adding the xpath from the other logout button and it worked.  THe
issue with using driver.quit still occurs but doesn't affect the test.
Something to do with streams?  See bug for further info